[FEAT] #64 프로필 이미지 변경 구현
